reorganiz user routes and add getOneUser Controller in User controller
