--- Routes/user.js
@@ -1,12 +1,0 @@
-
-const express = require("express");
-const router = express.Router();
-const UsersCtrl = require('../Controller/users');
-const TokenCtrl = require('../Controller/Auth')
-
-router.post('/', TokenCtrl);
-router.get('/', UsersCtrl.getAllUsers);
-router.post('/login', UsersCtrl.login);
-router.post('/register', UsersCtrl.register);
-
-module.exports = router;
